fixes from test
The build is in much better shape now: the library, examples, and benchmarks were all linking, and the only hard stop was test compilation. I patched the tests to avoid the Catch2 TEMPLATE_TEST_CASE_SIG ambiguity in key_value_store_test.hip.cpp and single_value_hash_table_test.hip.cpp, and I fixed the test-only unused-parameter warning there too. I also cleaned the %llu format warnings in multi_value_hash_table.hip.cpp, bucket_list_hash_table.hip.cpp, and multi_bucket_hash_table.hip.cpp.

+++ examples/src/bucket_list_hash_table.hip.cpp
@@ -192,7 +192,9 @@ int main ()
 
             if(upper - lower != size_values_per_key)
             {
-                printf("ERROR size values %llu\n", upper - lower);
+                printf(
+                    "ERROR size values %llu\n",
+                    static_cast<unsigned long long>(upper - lower));
             }
 
             for (index_t i = lower; i < upper; i++)
